+++ Sources/LCP/License/Container/LicenseContainer.swift
@@ -12,19 +12,13 @@ protocol LicenseContainer {
     /// Returns whether this container currently contains a License Document.
     ///
     /// For example, when fulfilling an EPUB publication, it initially doesn't contain the license.
-    func containsLicense() -> Bool
+    func containsLicense() async throws -> Bool
 
-    func read() throws -> Data
-    func write(_ license: LicenseDocument) throws
+    func read() async throws -> Data
+    func write(_ license: LicenseDocument) async throws
 }
 
-func makeLicenseContainer(for file: FileURL, mimetypes: [String] = []) -> Deferred<LicenseContainer?, LCPError> {
-    deferred(on: .global(qos: .background)) { success, _, _ in
-        success(makeLicenseContainerSync(for: file, mimetypes: mimetypes))
-    }
-}
-
-func makeLicenseContainerSync(for file: FileURL, mimetypes: [String] = []) -> LicenseContainer? {
+func makeLicenseContainer(for file: FileURL, mimetypes: [String] = []) -> LicenseContainer? {
     guard let mediaType = MediaType.of(file, mediaTypes: mimetypes, fileExtensions: []) else {
         return nil
     }