+++ Sources/LCP/Services/DeviceService.swift
@@ -8,7 +8,7 @@ import Foundation
 import ReadiumShared
 
 final class DeviceService {
-    private let repository: DeviceRepository
+    private let repository: LCPLicenseRepository
     private let httpClient: HTTPClient
 
     /// Returns the device's name.
@@ -16,7 +16,7 @@ final class DeviceService {
 
     init(
         deviceName: String,
-        repository: DeviceRepository,
+        repository: LCPLicenseRepository,
         httpClient: HTTPClient
     ) {
         name = deviceName
@@ -46,25 +46,25 @@ final class DeviceService {
     /// Registers the device for the given license.
     /// If the call was made, the updated Status Document data is given to the completion closure.
     @discardableResult
-    func registerLicense(_ license: LicenseDocument, at link: Link) -> Deferred<Data?, Error> {
-        deferredCatching {
-            let registered = try self.repository.isDeviceRegistered(for: license)
-            guard !registered else {
-                return .success(nil)
-            }
-            guard let url = link.url(parameters: self.asQueryParameters) else {
-                throw LCPError.licenseInteractionNotAvailable
-            }
-
-            return self.httpClient.fetch(HTTPRequest(url: url, method: .post))
-                .tryMap { response in
-                    guard 100 ..< 400 ~= response.statusCode else {
-                        return nil
-                    }
-
-                    try self.repository.registerDevice(for: license)
-                    return response.body
-                }
+    func registerLicense(_ license: LicenseDocument, at link: Link) async throws -> Data? {
+        let registered = try await repository.isDeviceRegistered(for: license.id)
+        guard !registered else {
+            return nil
         }
+        guard let url = link.url(parameters: asQueryParameters) else {
+            throw LCPError.licenseInteractionNotAvailable
+        }
+
+        let data = await httpClient.fetch(HTTPRequest(url: url, method: .post))
+            .map { response -> Data? in
+                guard 100 ..< 400 ~= response.statusCode else {
+                    return nil
+                }
+                return response.body
+            }
+
+        try await repository.registerDevice(for: license.id)
+
+        return try data.get()
     }
 }