+++ Sources/Navigator/EditingAction.swift
@@ -138,8 +138,6 @@ final class EditingActionsController {
     /// Verifies that the user has the rights to use the given `action`.
     private func isActionAllowed(_ action: EditingAction) -> Bool {
         switch action {
-        case .copy:
-            return rights.canCopy
         case .share:
             return canShare
         default:
@@ -188,11 +186,12 @@ final class EditingActionsController {
     }
 
     /// Copies the authorized portion of the selection text into the pasteboard.
-    func copy() {
+    @MainActor
+    func copy() async {
         guard let text = selection?.locator.text.highlight else {
             return
         }
-        guard rights.copy(text: text) else {
+        guard await rights.copy(text: text) else {
             delegate?.editingActionsDidPreventCopy(self)
             return
         }