+++ Sources/Shared/Fetcher/HTTPFetcher.swift
@@ -54,28 +54,52 @@ public final class HTTPFetcher: Fetcher, Loggable {
         }
 
         /// Cached HEAD response to get the expected content length and other metadata.
-        private lazy var headResponse: ResourceResult<HTTPResponse> = client.fetchSync(HTTPRequest(url: url, method: .head))
+        private lazy var headResponse: ResourceResult<HTTPResponse> = client.fetchWait(HTTPRequest(url: url, method: .head))
             .mapError { ResourceError.wrap($0) }
 
         /// An HTTP resource is always remote.
         var file: FileURL? { nil }
 
         func stream(range: Range<UInt64>?, consume: @escaping (Data) -> Void, completion: @escaping (ResourceResult<Void>) -> Void) -> Cancellable {
-            var request = HTTPRequest(url: url)
-            if let range = range {
-                request.setRange(range)
-            }
-
-            return client.stream(
-                request,
-                receiveResponse: nil,
-                consume: { data, _ in consume(data) },
-                completion: { result in
-                    completion(result.map { _ in }.mapError { ResourceError.wrap($0) })
+            let request = {
+                var request = HTTPRequest(url: url)
+                if let range = range {
+                    request.setRange(range)
                 }
-            )
+                return request
+            }()
+
+            return CancellableTask(task: Task {
+                let result = await client.stream(
+                    request: request,
+                    consume: { data, _ in consume(data) }
+                )
+                completion(result.map { _ in }.mapError { ResourceError.wrap($0) })
+            })
         }
 
         func close() {}
     }
+}
+
+private extension HTTPClient {
+    // FIXME: Get rid of this hack.
+    func fetchWait(_ request: HTTPRequestConvertible) -> HTTPResult<HTTPResponse> {
+        warnIfMainThread()
+
+        let enclosure = Enclosure()
+        let semaphore = DispatchSemaphore(value: 0)
+
+        Task {
+            enclosure.value = await fetch(request)
+            semaphore.signal()
+        }
+        _ = semaphore.wait(timeout: .distantFuture)
+
+        return enclosure.value
+    }
+}
+
+class Enclosure {
+    var value: HTTPResult<HTTPResponse>!
 }