+++ Sources/Shared/Toolkit/HTTP/HTTPClient.swift
@@ -15,53 +15,30 @@ public protocol HTTPClient: Loggable {
     ///
     /// - Parameters:
     ///   - request: Request to the streamed resource.
-    ///   - receiveResponse: Callback called when receiving the initial response, before consuming its body. You can
     ///     also access it in the completion block after consuming the data.
     ///   - consume: Callback called for each chunk of data received. Callers are responsible to accumulate the data
     ///     if needed.
-    ///   - completion: Callback called when the streaming finishes or an error occurs.
-    /// - Returns: A `Cancellable` interrupting the stream when requested.
     func stream(
-        _ request: HTTPRequestConvertible,
-        receiveResponse: ((HTTPResponse) -> Void)?,
-        consume: @escaping (_ chunk: Data, _ progress: Double?) -> Void,
-        completion: @escaping (HTTPResult<HTTPResponse>) -> Void
-    ) -> Cancellable
+        request: HTTPRequestConvertible,
+        consume: @escaping (_ chunk: Data, _ progress: Double?) -> Void
+    ) async -> HTTPResult<HTTPResponse>
 }
 
 public extension HTTPClient {
-    func stream(_ request: HTTPRequestConvertible, consume: @escaping (Data, Double?) -> Void, completion: @escaping (HTTPResult<HTTPResponse>) -> Void) -> Cancellable {
-        stream(request, receiveResponse: nil, consume: consume, completion: completion)
-    }
-
     /// Fetches the resource from the given `request`.
-    func fetch(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<HTTPResponse>) -> Void) -> Cancellable {
+    func fetch(_ request: HTTPRequestConvertible) async -> HTTPResult<HTTPResponse> {
         var data = Data()
-        return stream(request,
-                      consume: { chunk, _ in data.append(chunk) },
-                      completion: { result in
-                          completion(result.map {
-                              var response = $0
-                              response.body = data
-                              return response
-                          })
-                      })
-    }
+        let response = await stream(
+            request: request,
+            consume: { chunk, _ in data.append(chunk) }
+        )
 
-    /// Fetches a resource synchronously.
-    func fetchSync(_ request: HTTPRequestConvertible) -> HTTPResult<HTTPResponse> {
-        warnIfMainThread()
-
-        var result: HTTPResult<HTTPResponse>!
-
-        let semaphore = DispatchSemaphore(value: 0)
-        _ = fetch(request) {
-            result = $0
-            semaphore.signal()
-        }
-        _ = semaphore.wait(timeout: .distantFuture)
-
-        return result!
+        return response
+            .map {
+                var response = $0
+                response.body = data
+                return response
+            }
     }
 
     /// Fetches the resource and attempts to decode it with the given `decoder`.
@@ -69,45 +46,45 @@ public extension HTTPClient {
     /// If the decoder fails, a `malformedResponse` HTTP error is returned.
     func fetch<T>(
         _ request: HTTPRequestConvertible,
-        decoder: @escaping (HTTPResponse, Data) throws -> T?,
-        completion: @escaping (HTTPResult<T>) -> Void
-    ) -> Cancellable {
-        fetch(request) { response in
-            let result = response.flatMap { response -> HTTPResult<T> in
-                guard
-                    let body = response.body,
-                    let result = try? decoder(response, body)
-                else {
-                    return .failure(HTTPError(kind: .malformedResponse))
+        decoder: @escaping (HTTPResponse, Data) throws -> T?
+    ) async -> HTTPResult<T> {
+        await fetch(request)
+            .flatMap { response in
+                do {
+                    guard
+                        let body = response.body,
+                        let result = try decoder(response, body)
+                    else {
+                        return .failure(HTTPError(kind: .malformedResponse))
+                    }
+                    return .success(result)
+
+                } catch {
+                    return .failure(HTTPError(kind: .malformedResponse, cause: error))
                 }
-                return .success(result)
             }
-            completion(result)
-        }
     }
 
     /// Fetches the resource as a JSON object.
-    func fetchJSON(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<[String: Any]>) -> Void) -> Cancellable {
-        fetch(request,
-              decoder: { try JSONSerialization.jsonObject(with: $1) as? [String: Any] },
-              completion: completion)
+    func fetchJSON(_ request: HTTPRequestConvertible) async -> HTTPResult<[String: Any]> {
+        await fetch(request) {
+            try JSONSerialization.jsonObject(with: $1) as? [String: Any]
+        }
     }
 
     /// Fetches the resource as a `String`.
-    func fetchString(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<String>) -> Void) -> Cancellable {
-        fetch(request,
-              decoder: { response, body in
-                  let encoding = response.mediaType.encoding ?? .utf8
-                  return String(data: body, encoding: encoding)
-              },
-              completion: completion)
+    func fetchString(_ request: HTTPRequestConvertible) async -> HTTPResult<String> {
+        await fetch(request) { response, body in
+            let encoding = response.mediaType.encoding ?? .utf8
+            return String(data: body, encoding: encoding)
+        }
     }
 
     /// Fetches the resource as an `UIImage`.
-    func fetchImage(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<UIImage>) -> Void) -> Cancellable {
-        fetch(request,
-              decoder: { UIImage(data: $1) },
-              completion: completion)
+    func fetchImage(_ request: HTTPRequestConvertible) async -> HTTPResult<UIImage> {
+        await fetch(request) {
+            UIImage(data: $1)
+        }
     }
 
     /// Downloads the resource at a temporary location.
@@ -115,9 +92,8 @@ public extension HTTPClient {
     /// You are responsible for moving or deleting the downloaded file in the `completion` block.
     func download(
         _ request: HTTPRequestConvertible,
-        onProgress: @escaping (Double) -> Void,
-        completion: @escaping (HTTPResult<HTTPDownload>) -> Void
-    ) -> Cancellable {
+        onProgress: @escaping (Double) -> Void
+    ) async -> HTTPResult<HTTPDownload> {
         let location = FileURL(
             url: URL(
                 fileURLWithPath: NSTemporaryDirectory(),
@@ -130,16 +106,11 @@ public extension HTTPClient {
             try "".write(to: location.url, atomically: true, encoding: .utf8)
             fileHandle = try FileHandle(forWritingTo: location.url)
         } catch {
-            completion(.failure(HTTPError(kind: .ioError, cause: error)))
-            return CancellableObject()
+            return .failure(HTTPError(kind: .ioError, cause: error))
         }
-        var suggestedFilename: String?
 
-        return stream(
-            request,
-            receiveResponse: { response in
-                suggestedFilename = response.filename
-            },
+        let result = await stream(
+            request: request,
             consume: { data, progression in
                 fileHandle.seekToEndOfFile()
                 fileHandle.write(data)
@@ -147,34 +118,90 @@ public extension HTTPClient {
                 if let progression = progression {
                     onProgress(progression)
                 }
-            },
-            completion: { result in
-                if #available(iOS 13.0, *) {
-                    do {
-                        try fileHandle.close()
-                    } catch {
-                        log(.warning, error)
-                    }
-                }
-
-                switch result {
-                case let .success(response):
-                    completion(.success(HTTPDownload(
-                        location: location,
-                        suggestedFilename: suggestedFilename ?? response.filename,
-                        mediaType: response.mediaType
-                    )))
-
-                case let .failure(error):
-                    completion(.failure(error))
-                    do {
-                        try FileManager.default.removeItem(at: location.url)
-                    } catch {
-                        log(.warning, error)
-                    }
-                }
             }
         )
+
+        do {
+            try fileHandle.close()
+        } catch {
+            log(.warning, error)
+        }
+
+        switch result {
+        case let .success(response):
+            return .success(HTTPDownload(
+                location: location,
+                suggestedFilename: response.filename,
+                mediaType: response.mediaType
+            ))
+
+        case let .failure(error):
+            do {
+                try FileManager.default.removeItem(at: location.url)
+            } catch {
+                log(.warning, error)
+            }
+            return .failure(error)
+        }
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func stream(_ request: HTTPRequestConvertible, receiveResponse: ((HTTPResponse) -> Void)?, consume: @escaping (_ chunk: Data, _ progress: Double?) -> Void, completion: @escaping (HTTPResult<HTTPResponse>) -> Void) -> Cancellable {
+        CancellableTask(task: Task {
+            await completion(stream(request: request, consume: consume))
+        })
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func stream(_ request: HTTPRequestConvertible, consume: @escaping (Data, Double?) -> Void, completion: @escaping (HTTPResult<HTTPResponse>) -> Void) -> Cancellable {
+        CancellableTask(task: Task {
+            await completion(stream(request: request, consume: consume))
+        })
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetch(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<HTTPResponse>) -> Void) -> Cancellable {
+        CancellableTask(task: Task {
+            await completion(fetch(request))
+        })
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetchSync(_ request: HTTPRequestConvertible) -> HTTPResult<HTTPResponse> {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetch<T>(
+        _ request: HTTPRequestConvertible,
+        decoder: @escaping (HTTPResponse, Data) throws -> T?,
+        completion: @escaping (HTTPResult<T>) -> Void
+    ) -> Cancellable {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetchJSON(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<[String: Any]>) -> Void) -> Cancellable {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetchString(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<String>) -> Void) -> Cancellable {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func fetchImage(_ request: HTTPRequestConvertible, completion: @escaping (HTTPResult<UIImage>) -> Void) -> Cancellable {
+        fatalError()
+    }
+
+    @available(*, unavailable, message: "Use the async variant.")
+    func download(
+        _ request: HTTPRequestConvertible,
+        onProgress: @escaping (Double) -> Void,
+        completion: @escaping (HTTPResult<HTTPDownload>) -> Void
+    ) -> Cancellable {
+        fatalError()
     }
 }