+++ TestApp/Sources/Library/LibraryViewController.swift
@@ -212,7 +212,7 @@ extension LibraryViewController: UICollectionViewDelegateFlowLayout, UICollectio
         // Load image and then apply the shadow.
         if
             let coverURL = book.cover,
-            let data = try? Data(contentsOf: coverURL),
+            let data = try? Data(contentsOf: coverURL.url),
             let cover = UIImage(data: data)
         {
             cell.coverImageView.image = cover